+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Pin `moment` at `~2.24.0` in light of multiple issues with `2.25.0` ([5489](https://github.com/moment/moment/issues/5489), [5472](https://github.com/moment/moment/issues/5472)).
 * Purge `intlShape` in prep for `react-intl` `v4` migration. Refs STRIPES-672.
 * Add ability to search by `Effective call number (item), normalized` option. Refs UIIN-993.
+* Add ability to search by `Call number, normalized` option via holdings segment. Refs UIIN-983.
 
 ## [2.0.0](https://github.com/folio-org/ui-inventory/tree/v2.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.1...v2.0.0)